Introduce the WearableConnectionRepository
