+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/ui/login/LoginViewModel.kt
@@ -4,6 +4,8 @@ import android.os.Parcelable
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.asLiveData
 import androidx.navigation.NavHostController
+import com.woocommerce.android.sync.WearableConnectionRepository
+import com.woocommerce.android.sync.WearableConnectionRepository.MessagePath.START_AUTH
 import com.woocommerce.android.ui.NavRoutes.MY_STORE
 import com.woocommerce.android.viewmodel.ScopedViewModel
 import com.woocommerce.android.viewmodel.getStateFlow
@@ -18,6 +20,7 @@ import kotlinx.parcelize.Parcelize
 @HiltViewModel(assistedFactory = LoginViewModel.Factory::class)
 class LoginViewModel @AssistedInject constructor(
     private val loginRepository: LoginRepository,
+    private val connRepository: WearableConnectionRepository,
     @Assisted private val navController: NavHostController,
     savedState: SavedStateHandle
 ) : ScopedViewModel(savedState) {
@@ -44,7 +47,7 @@ class LoginViewModel @AssistedInject constructor(
     fun onLoginButtonClicked() {
         _viewState.update { it.copy(isLoading = true) }
         launch {
-            loginRepository.sendMessageToAllNodes("/start-auth")
+            connRepository.sendMessageToAllNodes(START_AUTH)
             _viewState.update { it.copy(isLoading = false) }
         }
     }
